Assignment

+++ app/src/main/java/com/example/demo/model/TaskModel.kt
@@ -6,7 +6,7 @@ import androidx.room.*
 data class TaskModel(
 
     @PrimaryKey(autoGenerate = true)
-    val taskModelId: Int,
+    val taskModelId: Long,
 
     val date: String? = null,
